// src/stories/TextArea/styles.module.scss.d.ts
declare const styles:
  & Readonly<{ "textArea": string }>
  & Readonly<{ "textArea": string }>
  & Readonly<{ "textArea": string }>
  & Readonly<{ "isInvalid": string }>
  & Readonly<{ "wrapper": string }>
  & Readonly<{ "label": string }>
  & Readonly<{ "error": string }>
  & Readonly<{ "characterCount": string }>
  & Readonly<{ "labelAndCountContainer": string }>
  & Readonly<{ "requiredApostrophe": string }>
;
export default styles;
//# sourceMappingURL=./styles.module.scss.d.ts.map